cli: move the quit code to its own file
Release note: None
